\name{BrowserViewList-class}
\docType{class}
\alias{BrowserViewList-class}

% Constructor
\alias{BrowserViewList}

\title{Lists of BrowserView} 

\description{ A formal list of
  \code{\linkS4class{BrowserView}} objects. Extends and inherits all
  its methods from
  \code{\link[S4Vectors]{Vector}}. Usually generated by
  passing multiple ranges to the \code{\link{browserView}} function.
}

\section{Constructor}{
  \describe{
  \item{\code{BrowserViewList(...)}}{Concatenates the
      \code{BrowserView} objects in \code{...} into a new
      \code{BrowserViewList}.  This is rarely called by the user.
    } 
  } 
}

\author{Michael Lawrence}

\keyword{classes}
