\name{connection-class}
\alias{connection-class}
\alias{characterORconnection-class}
\alias{file-class}
\alias{gzfile-class}
\alias{pipe-class}
\alias{sockconn-class}
\alias{unz-class}
\alias{url-class}
\alias{terminal-class}
\alias{textConnection-class}

\title{S4 connection classes}
\description{ These are S4 representations of the S3 connection classes
  in R. They exist only to support method dispatch. }
\author{ Michael Lawrence }
\keyword{classes}
\keyword{internal}