\name{laySequence-methods}
\docType{methods}
\alias{laySequence}
\alias{laySequence-methods}
\title{ Load a sequence }
\description{
  Methods for loading sequences.
}
\section{Methods}{
\describe{
  No methods are defined by \pkg{rtracklayer} for the
  \code{laySequence(object, sequence, ...)} generic.
}}
\keyword{methods}
