\name{ucscGenomes}
\alias{ucscGenomes}
\title{Get available genomes on UCSC}
\description{
  Get a list of genome identifiers for the UCSC genome browser.
}
\usage{
ucscGenomes()
}
\value{
  A character vector of genome identifiers, with names indicating the species.
}
\seealso{\code{\linkS4class{UCSCSession}} for details on specifying the
  genome.}
\author{ Michael Lawrence }
\keyword{interface}
