\name{XMFAFile-class}
\docType{class}

%% Classes:
\alias{class:XMFAFile}
\alias{XMFAFile-class}

%% Constructor:
\alias{XMFAFile}

%% Methods:
\alias{import,XMFAFile,ANY,ANY-method}

\title{XMFAFile objects}

\description{
  Imports the XMFA format generated by aligners like Mauve.
}

\usage{
\S4method{import}{XMFAFile,ANY,ANY}(con, format, text)
}

\arguments{
  \item{con}{A path, URL, connection or \code{XMFAFile} object
    (constructed by calling \code{XMFAFile(path)}).  If \code{con} is a
    path, URL or connection, either the file extension or the
    \code{format} argument needs to be \dQuote{xmfa}. Compressed files
    (\dQuote{gz}, \dQuote{bz2} and \dQuote{xz}) are handled
    transparently.
  }
  \item{format}{If not missing, should be \dQuote{xmfa}.
  }
  \item{text}{If \code{con} is missing, a character vector to use as the
    input.
  }
}

\value{
  An \code{\link[Biostrings]{AlignedXStringSetList}} object.
}

\details{
  The eXtended Multi-FastA (XMFA) format is a standard for communicating
  genome alignments with rearrangements. It consists of FASTA records
  representing alignments, each with a header indicating the genomic
  position of the alignment and whether there was an inversion
  (alignment to the minus strand). The popular Mauve aligner generates
  XMFA files.
}

\author{Michael Lawrence}

\references{
  \url{http://darlinglab.org/mauve/user-guide/files.html}
}

