\name{sequence<--methods}
\docType{methods}
\alias{sequence<-}
\alias{sequence<--methods}
\title{ Load a sequence }
\description{
  Methods for loading sequences.
}
\section{Methods}{
\describe{
  No methods are defined by \pkg{rtracklayer} for the
  \code{sequence(object, ...) <- value} generic.
}}
\keyword{methods}
